+++ test/structural_transformation/index_reduction.jl
@@ -115,8 +115,10 @@ prob_auto = ODEProblem(new_sys, u0, (0.0, 10.0), p)
 sol = solve(prob_auto, Rodas5());
 #plot(sol, idxs=(D(x), y))
 
-let pss_pendulum2 = partial_state_selection(pendulum2)
-    @test length(equations(pss_pendulum2)) <= 6
+@test_skip begin
+    let pss_pendulum2 = partial_state_selection(pendulum2)
+        length(equations(pss_pendulum2)) <= 6
+    end
 end
 
 eqs = [D(x) ~ w,